修复BigInteger

+++ src/algorithms/extsort/RecordStore.java
@@ -4,6 +4,7 @@
 package algorithms.extsort;
 
 import java.io.IOException;
+import java.math.BigInteger;
 
 /**
  * Should provide buffer mechanism
@@ -14,7 +15,7 @@ import java.io.IOException;
 public interface RecordStore
 {
 	Record readNextRecord() throws IOException;
-	Record readNextRecord(long index) throws IOException;
+	Record readNextRecord(BigInteger index) throws IOException;
 	void destroy();
 
 }
